tests/pyproject.toml
[project]
name = "egregore-tests"
version = "0.1.0"
description = "Testing framework for Egregore based on unit economics"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.0.0",
    "pytest-html>=4.0.0",
    "python-dotenv>=1.0.0",
    "neo4j>=5.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "capture: Capture reliability tests (p - items enter system)",
    "sync: Sync accuracy tests (p - items findable)",
    "retrieval: Retrieval stability tests (p × V - trust)",
    "quality: Data quality tests (p × V - answer usefulness)",
    "security: Security tests (C - incident cost)",
    "api: API endpoint tests",
    "isolation: Cross-tenant isolation tests",
    "flow: End-to-end flow tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[dependency-groups]
dev = []
